mysite/requirements.txt
appdirs==1.4.3
boto==2.48.0
boto3==1.4.4
certifi==2017.4.17
chardet==3.0.4
Django==1.11.1
django-common==0.1.51
django-cors-headers==2.1.0
django-extensions==1.7.9
django-pipeline==1.6.13
django-pipeline-browserify==0.6.1
django-shell-plus==1.1.7
django-storages==1.6.3
django-hashid-field==1.2.2
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
flake8==3.3.0
futures==3.1.1
idna==2.5
packaging==16.8
pprint==0.1
pyparsing==2.2.0
pytz==2017.2
requests==2.17.3
runscript==0.2.6
six==1.10.0
Unidecode==0.4.20
urllib3==1.21.1